# backend/src/main/java/com/diploma/mrt/model/FindingLocation.java
package com.diploma.mrt.model;

import java.util.List;

public record FindingLocation(
        String segment,
        List<Double> centroid,
        BoundingBox bbox,
        List<Integer> extent,
        String suspicion
) {
}