# ml-service/app/schemas/health_response.py
from pydantic import BaseModel


class HealthResponse(BaseModel):
    status: str
    mode: str
    version: str
    liverModelConfigured: bool
    lesionModelConfigured: bool
    mriHeuristicSupported: bool